developHotFix

+++ client/src/Redux/Actions/actions.js
@@ -20,7 +20,7 @@ import {
   PUT_ORDERS,
   PUT_ORDERS_AMOUNT,
   DELETE_SALE,
-  POST_NEW_ADRESS_USER
+  POST_NEW_ADRESS_USER,
 } from './types';
 
 requestInterceptor();
@@ -183,11 +183,9 @@ export function validateMail(mail) {
 
 export function postOrder(order) {
   const token = window.localStorage.getItem('access');
-  console.log(token);
   if (token) {
     return async function (dispatch) {
       var json = await axios.post(`http://localhost:3001/order`, order);
-      console.log(json.data);
       return dispatch({
         type: POST_ORDERS,
         payload: { status: order.status, data: json.data },
@@ -195,7 +193,6 @@ export function postOrder(order) {
     };
   } else {
     if (!window.localStorage.getItem(`${order.status}`)) {
-      console.log('hola2');
       const product = {
         status: order.status,
         amount: order.amount,
@@ -213,7 +210,6 @@ export function postOrder(order) {
       arr.push(product);
       window.localStorage.setItem(`${order.status}`, JSON.stringify(arr));
       var item = window.localStorage.getItem(`${order.status}`);
-      console.log(JSON.parse(item));
     } else {
       const product = {
         status: order.status,
@@ -291,7 +287,6 @@ export function getOrder(order) {
   if (token) {
     return async function (dispatch) {
       var json = await axios.get(`http://localhost:3001/order?status=` + order.status);
-      console.log(json.data);
       return dispatch({
         type: GET_ORDERS,
         payload: { status: order.status, data: json.data },
@@ -301,7 +296,6 @@ export function getOrder(order) {
     return function (dispatch) {
       const item = window.localStorage.getItem(`${order.status}`);
       const parsedItem = JSON.parse(item);
-      console.log(parsedItem);
       return dispatch({
         type: GET_ORDERS,
         payload: { status: order.status, data: parsedItem },
@@ -321,7 +315,6 @@ export function changeOrderStatus(order) {
 }
 
 export function changeOrderAmount(order) {
-  console.log(order);
   return async function (dispatch) {
     var json = await axios.put(`/order/${order.id}?amount=${order.amount}`);
     return dispatch({
@@ -395,14 +388,31 @@ export function changeOrderAmount(order) {
    /*    const token = window.localStorage.getItem('access')
   } */
 
-export function deleteOrder(order) {
-  return async function (dispatch) {
-    var json = await axios.delete(`/order/${order}`);
-    return dispatch({
-      type: DELETE_ORDERS,
-      payload: json.data,
-    });
-  };
+export function deleteOrder(order, id, status) {
+  const token = window.localStorage.getItem('access');
+  if (token) {
+    return async function (dispatch) {
+      var json = await axios.delete(`/order/${order}`);
+      return dispatch({
+        type: DELETE_ORDERS,
+        payload: json.data,
+      });
+    };
+  } else {
+    return function (dispatch) {
+      console.log(id);
+      const item = window.localStorage.getItem(`${status}`);
+      const parsedItem = item && JSON.parse(item);
+      console.log(parsedItem);
+      const itemDeleted = parsedItem && parsedItem.filter((el) => el.productId !== id);
+      console.log(itemDeleted);
+      window.localStorage.setItem(`${status}`, JSON.stringify(itemDeleted));
+      return dispatch({
+        type: DELETE_ORDERS,
+        payload: { status: status, data: itemDeleted },
+      });
+    };
+  }
 }
 /*    const token = window.localStorage.getItem('access')
 
@@ -514,21 +524,18 @@ export function putUserInfo(token, body) {
 }
 
 export function postDirectionUser(body) {
-  console.log(body);
   return async (dispatch) => {
     try {
-      const newAdress = await axios.post("/user/direction", body);
+      const newAdress = await axios.post('/user/direction', body);
       return dispatch({
         type: POST_NEW_ADRESS_USER,
         payload: newAdress.data,
       });
     } catch (error) {
-      console.log(error)
+      console.log(error);
     }
-  }
+  };
 }
-
-
 
 export function getSales() {
   return async function (dispatch) {
